requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
pypdf>=4.0.0
sentence-transformers>=2.7.0
python-dotenv>=1.0.0
PyYAML>=6.0.0